requirements.txt
pyyaml
imagesize
